refactor: Refactor routing and page organization
Move `Chat` and `Profile` pages out of `/Main` directory.
Remove unnecessary `Progress` route from main navigation.  Improve
code formatting and structure for better readability.
